completed all timers. need verification + AVI fix
AVI needs to give priority to URI. That is like in case (4) of the diagram where the fastest time dictated by URI is violated by AVI, so AVI gets shafted. Easiest way to fix is in the AVI macrostate, the priority 2 arc for waitAVI --> intState should add a "&& URI_TO  = true" to ensure the upper rate interval, URI, has completed before actually asserting a VPace. I.e. from the URI macrostate, it should be in its Idle state. Idk how to do that though, or if it should be a C thing